Refactored the fuck out of this code
This code was fucked. So much repeated and unecessary code and now look 
at where we are. 

The player class is now it's own actual class and all of it's related 
functions are in there.

The player class returns the player sprite you give it, and you must 
call player.functions to use it's methods

Also made bases Nidhogg style.
problem when both kill eachother at the same time because player2's 
function runs second