Done refactoring, added Base class
Base class is now a field in the Player class, this way we can deal with 
bases as connected to a player. 

Basically done refactoring the code now, everything is in it's classes.

If we could actually create the animations in the player class if we 
generalize it, but it's not a big deal and that is not very necessary, 
it would only just move that code to the player class which could be 
nice but is whatever.